requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
async-timeout==5.0.1
click==8.1.8
Django==4.2.25
django-environ==0.12.0
djangorestframework==3.16.1
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.9
pydantic==2.11.9
pydantic_core==2.33.2
redis==6.4.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
gunicorn
websockets>=10.0

